Fix IE <= 8
If I apply pixrem to this css:

`a { font-size: 0.75rem; }`

I would expect a fallback to be added for IE <= 8, but not IE >= 9.  

If I specify 'ie 8' in the browser option then everything works ok, but if I specify 'ie >= 8' then the fallback is not added, but I think it should be.

I think line 76 in pixrem.js should change from:

`var isNotUseless = (!isIE9_10 && !isUseless);`

to:

`var isNotUseless = (isIElte8 && !isUseless);`

As it is now, all rules are considered useless if IE9 or 10 are included in the browser list.